+++ src/connectedk8s/HISTORY.rst
@@ -3,6 +3,15 @@
 Release History
 ===============
 
+1.1.10
+++++++
+
+* Fixed ARM exception telemetry
+
+1.1.9
+++++++
+
+* Increase onboarding and upgrade timeout
 
 1.1.8
 ++++++